refactor(syntax-refactor): refactor of jurisdiction name
